[feat] : 히스토리 경로 변경 (#56)
