+++ lib/view/my_page/history_vege_screen.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:mojacknong_android/common/farmus_theme_data.dart';
 import 'package:mojacknong_android/common/primary_app_bar.dart';
 
-class MyPageList extends StatelessWidget {
+class HistoryVegeScreen extends StatelessWidget {
   final String? data;
 
-  const MyPageList({Key? key, required this.data}) : super(key: key);
+  const HistoryVegeScreen({Key? key, required this.data}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -12,6 +13,7 @@ class MyPageList extends StatelessWidget {
       appBar: PrimaryAppBar(
         title: "채소 히스토리",
       ),
+      backgroundColor: FarmusThemeData.white,
       body: Center(
         child: Text('New page with data: $data'),
       ),
